requirements.txt
# ── Core Framework ──────────────────────────────────────────
streamlit==1.32.0

# ── LangChain & Integrations ────────────────────────────────
langchain==0.1.13
langchain-community==0.0.29
langchain-openai==0.1.1
langchain-google-genai==1.0.1

# ── Vector Database ──────────────────────────────────────────
chromadb==0.4.24

# ── Document Loaders ─────────────────────────────────────────
pypdf==4.1.0
python-docx==1.1.0

# ── LLM Providers ────────────────────────────────────────────
openai==1.14.3
google-generativeai==0.4.1

# ── Embeddings ───────────────────────────────────────────────
tiktoken==0.6.0

# ── Utilities ────────────────────────────────────────────────
python-dotenv==1.0.1
loguru==0.7.2
pytest==8.1.1
